+++ dimcli/core/api.py
@@ -233,6 +233,31 @@ class Result(IPython.display.JSON):
     >>> res['stats'] # => the _stats dict
 
     """
+
+    # class methods to build Result object from raw data (obtained not from a query eg from multiple queries concatenated)
+    # these allow to then take advantage of other functionalities in Result objects eg dataframes etc...
+    @classmethod
+    def from_publications_list(cls, data):
+        return cls({"publications" : data, '_stats' : {'total_count' : len(data)}})
+    @classmethod
+    def from_grants_list(cls, data):
+        return cls({"grants" : data, '_stats' : {'total_count' : len(data)}})
+    @classmethod
+    def from_researchers_list(cls, data):
+        return cls({"researchers" : data, '_stats' : {'total_count' : len(data)}})
+    @classmethod
+    def from_clinical_trials_list(cls, data):
+        return cls({"clinical_trials" : data, '_stats' : {'total_count' : len(data)}})
+    @classmethod
+    def from_patents_list(cls, data):
+        return cls({"patents" : data, '_stats' : {'total_count' : len(data)}})
+    @classmethod
+    def from_policy_documents_list(cls, data):
+        return cls({"policy_documents" : data, '_stats' : {'total_count' : len(data)}})
+    @classmethod
+    def from_organizations_list(cls, data):
+        return cls({"organizations" : data, '_stats' : {'total_count' : len(data)}})
+
     def __init__(self, data):
         IPython.display.JSON.__init__(self, data)
         self.json = self.data